init: create react app and google maps api
The application is started from boilerplate a react template using the initializer react-app made by facebook. So, the file structure created the main react files, an html file with the root element and a javascript file that controls it.
The Google maps API needs the initMap function in the main app component. Also, in order to create a script element in the template html a load script function is created that collects the first script element. Then, a new script element is created and interted it in the collected element to the parent, so that to be the first. The map that fetched from the Google maps brings the map of Athens historical center at a high zoom to show a neighborhood of ancient Athens. A key provided by Google is used to the account of the developer.