feat: get current location data for food venues
The application uses Foursquare API for getting venues recomendations about the current map that is Athens historical center. In order to fetch data from API used axios an HTTP client using foursquare account keys of the developer.
To invoke the implemented method of getting venues it is used by the component-did-mount lifecycle method of React. To make the fetched data usable to other components, a state of venues array is implemented.