feat: add marker for every venue fetched
Using the Google map markers creating markers for every venue fetched from Foursquare venue explore API. In order to avoid rendering map before getting venues from API used promise function, so that first to perform the asyncronous request of fetching and then put venuses to state, so to storing them for further use. Last of this procedure is to render map on screen with markers of the venues. Also, a title of the venue is popping when mouse is pointing on marker.