feat: display infowindow on marker click
Info-windows is Google mas API method for creating information boxes on screen that shows interesting text and images related to the spot on map that is attatched. Every info-window is popup above the map at a given location. In this application one info-window is created and on click of a marker a box with content is setting an info-window.