feat: App Structure and List of Places component implementation
The application have some file structure, so that, the data files to be in a place with some app dafaults to store license keys. A new folder place for the utils, in means of useful methods is made to the app structure.
A new component List of places made for the side list, with the names of sights in the neighborhood of the app to visit.A new component of map made to render the map on screen. Also, some improvements made in the app logic made to the main App component.