feat: Added functionality of Info-window and Search place
The Info-window on marker is implemented so that, when clicked on a marker to pop-up info about the location. Search functionality is implemented so that, when starting typing in the search form simultanusly the list of locations is filtering the results and the markers display on the map accordingly the result of the filtering. The places on the list of places are button elements, that can open info-wiki box with further info about the place.

This is an extended commit with app file structure change and other changes also made. Improve appearence of the app must be done in the nexts commits with CSS fixes.