fix: after Code Review issues about Functionality & Error msgs
Fix issue to let the location items that appear in list when get clicked go to the actual location. In order to fix this set center method of Google maps used to recenter map in the browser window.
Fix issue have to do with error messages to appear when no internet connection.